Add files via upload
GROUP 2 / alternative approach to building text-to-speech system.

Phoneset, Lexicon and Voice-Setup are defined each in a different scm-file,
hhu_phones.scm, hhulex-scm, hhu_diphone.scm respectively.

lexicons.scm is an already existing file in festival. It is uploaded here for viewing purposes, because a few lines of code were added.